+++ doc/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   Open recent files. (#527)
 -   Reveal the executable file in file manager. (#537)
+-   Use a dialog to open new contests. (#539 and #467)
 
 ### Fixed